requirements.txt
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular[sidecar]>=0.27
django-cors-headers>=4.4